network_topology_w_rf_coil/backend_integration/scanner_integration_service.py
import os
import time
import json
import logging
from abc import ABC, abstractmethod

# Mocking libraries for demonstration if they aren't installed
try:
    import boto3
except ImportError:
    boto3 = None

try:
    from google.cloud import storage
except ImportError:
    storage = None

# Configure logging
logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(name)s - %(levelname)s - %(message)s')
logger = logging.getLogger("SunnybrookIntegration")

class ImagingService(ABC):
    @abstractmethod
    def upload_image(self, file_path, metadata):
        pass

    @abstractmethod
    def process_image(self, image_id, analysis_type):
        pass

class AWSSageMakerService(ImagingService):
    def __init__(self, bucket_name, region="ca-central-1"):
        self.bucket_name = bucket_name
        self.region = region
        if boto3:
            self.s3 = boto3.client('s3', region_name=region)
            self.sm_runtime = boto3.client('sagemaker-runtime', region_name=region)
        else:
            logger.warning("boto3 not installed, running in mock mode for AWS")
            self.s3 = None
            self.sm_runtime = None

    def upload_image(self, file_path, metadata):
        file_name = os.path.basename(file_path)
        if self.s3:
            self.s3.upload_file(file_path, self.bucket_name, file_name, ExtraArgs={"Metadata": metadata})
            logger.info(f"Uploaded {file_name} to AWS S3 bucket {self.bucket_name}")
        else:
            logger.info(f"[MOCK] Uploaded {file_name} to AWS S3 bucket {self.bucket_name} with metadata {metadata}")
        return f"s3://{self.bucket_name}/{file_name}"

    def process_image(self, image_id, analysis_type):
        endpoint_name = f"sunnybrook-{analysis_type}-endpoint"
        logger.info(f"Invoking SageMaker endpoint {endpoint_name} for image {image_id}")
        # In a real scenario, we would payload the image data 
        return {"status": "processing", "job_id": f"aws-{int(time.time())}"}

class GCPVertexService(ImagingService):
    def __init__(self, bucket_name, project_id):
        self.bucket_name = bucket_name
        self.project_id = project_id
        if storage:
            self.client = storage.Client(project=project_id)
        else:
            logger.warning("google-cloud-storage not installed, running in mock mode for GCP")
            self.client = None

    def upload_image(self, file_path, metadata):
        file_name = os.path.basename(file_path)
        if self.client:
            bucket = self.client.bucket(self.bucket_name)
            blob = bucket.blob(file_name)
            blob.metadata = metadata
            blob.upload_from_filename(file_path)
            logger.info(f"Uploaded {file_name} to GCP Bucket {self.bucket_name}")
        else:
            logger.info(f"[MOCK] Uploaded {file_name} to GCP Bucket {self.bucket_name}")
        return f"gs://{self.bucket_name}/{file_name}"

    def process_image(self, image_id, analysis_type):
        logger.info(f"Triggering Vertex AI pipeline for {analysis_type} on {image_id}")
        return {"status": "processing", "job_id": f"gcp-{int(time.time())}"}

class NvidiaQuantumSimulator:
    """
    Interface for experimental Quantum Image Processing.
    """
    def __init__(self):
        logger.info("Initializing NVIDIA Quantum (CuQuantum) Interface")

    def run_quantum_edge_detection(self, image_data):
        """
        Simulate a quantum edge detection algorithm.
        """
        logger.info("Encoding image into quantum state...")
        time.sleep(1) # Simulate processing
        logger.info("Applying QFT (Quantum Fourier Transform)...")
        time.sleep(1)
        logger.info("Collapsing state to classical output...")
        return "quantum_enhanced_edge_map.png"

class ScannerInterface:
    def __init__(self, scanner_ip, modality):
        self.scanner_ip = scanner_ip
        self.modality = modality

    def fetch_latest_study(self):
        logger.info(f"Connecting to {self.modality} scanner at {self.scanner_ip}...")
        # DICOM C-MOVE logic would go here
        return __file__ # Use current file as mock study to ensure existence

class HeadCoilDesignerService:
    """ Service to design patient-specific RF Head Coils using NVQLink. """
    def __init__(self):
        logger.info("Initializing NVQLink Head Coil Designer Service")

    def design_coil_for_scanner(self, field_strength_tesla: float):
        try:
            # Dynamic import to avoid circular dependencies if any
            import sys, os
            sys.path.append(os.path.join(os.path.dirname(__file__), '../rf_coils'))
            from nvqlink_head_coil import NVQLinkHeadCoilOptimizer
            
            optimizer = NVQLinkHeadCoilOptimizer(target_field_strength=field_strength_tesla)
            topology = optimizer.generate_optimal_topology()
            circuit = optimizer.analyze_circuitry(topology)
            
            logger.info(f"Generated Optimal Coil: {topology.turns} turns, Q={circuit.quality_factor:.2f}")
            return topology, circuit
        except Exception as e:
            logger.error(f"Failed to design coil: {e}")
            import traceback
            traceback.print_exc()
            return None, None

# Example Usage
if __name__ == "__main__":
    # Initialize services
    aws_service = AWSSageMakerService(bucket_name="sunnybrook-imaging-research-data-lake")
    gcp_service = GCPVertexService(bucket_name="sunnybrook-pathology-backup", project_id="sunnybrook-research")
    quantum_sim = NvidiaQuantumSimulator()
    mri_scanner = ScannerInterface("192.168.1.100", "MRI")
    coil_service = HeadCoilDesignerService()

    # Workflow Simulation
    logger.info("--- Starting Neuroimaging Workflow ---")
    
    # 1. Acquire Data
    study_path = mri_scanner.fetch_latest_study()
    
    # 2. Upload to Cloud (Hybrid)
    s3_uri = "s3://mock-bucket/mock-key"
    try:
        s3_uri = aws_service.upload_image(study_path, {"patient_id": "anonymized_001", "modality": "MRI"})
    except Exception as e:
        logger.warning(f"AWS Upload failed (expected in disconnected env): {e}")
    
    # 3. Process
    result = aws_service.process_image(s3_uri, "neuro-segmentation")
    logger.info(f"Analysis started: {result}")

    # 4. Experimental Quantum Processing
    experiment_result = quantum_sim.run_quantum_edge_detection("raw_data_matrix")
    logger.info(f"Quantum result: {experiment_result}")

    # 5. Coil Optimization (Pre-scan or Hardware Upgrade)
    logger.info("--- Starting NVQLink Coil Optimization ---")
    coil_service.design_coil_for_scanner(field_strength_tesla=3.0)
