network_topology_w_rf_coil/backend_integration/requirements.txt
boto3>=1.26.0
google-cloud-storage>=2.0.0
numpy>=1.21.0
pydicom>=2.3.0
